move file to correct package
