+++ src/java/Controller/home/ContactServlet.java
@@ -1,6 +1,6 @@
 package Controller.Home;
 
-import Constant.OAuthConstants;
+import Model.Constants.OAuthConstants;
 import java.io.IOException;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
